requirements.txt
matplotlib==3.3.4
numpy==1.19.2
pillow==8.1.2
scikit-image==0.18.1
scipy==1.6.1
torchvision==0.9.0
tqdm==4.59.0
albumentations~=0.5.2
